+++ GitHub/GitHubTests/AutoGenerated/OutputMocks.swift
@@ -1,0 +1,28 @@
+///
+/// @Generated by Mockolo
+///
+
+
+
+import Alamofire
+import Foundation
+import RealmSwift
+import UIKit
+@testable import GitHub
+
+
+class NewUsersRepositoryMock: NewUsersRepository {
+    init() { }
+
+
+    private(set) var requestUsersCallCount = 0
+    var requestUsersHandler: ((Int, @escaping (_ models: [UserModel]) -> Void) -> ())?
+    func requestUsers(userId: Int, completionHandler: @escaping (_ models: [UserModel]) -> Void)  {
+        requestUsersCallCount += 1
+        if let requestUsersHandler = requestUsersHandler {
+            requestUsersHandler(userId, completionHandler)
+        }
+        
+    }
+}
+
